moved summary generation into the main BIDS loop to avoid iterated the dataset twice
